Add generated JSON for Clarity's code
This is a hack to generate the JSON for Clarity's documentation.
We cloned their repo into a folder called `clarity-src` and ran the
newly added `build-clarity-doc` npm target (which is not referenced)
anywhere and checked in the generated JSON.

This will need to be redone as we need new versions of Clarity. This
could be improved to automate the process of cloning Clarity's repo,
running the script and deleting Clarity's code after the generated file
has been created

Signed-off-by: Juan Mendes <jmendes@vmware.com>